refactor(metrics): move serial device metrics out of logger
Move serial device metrics out of logger to devices.
Flush serial device metrics through legacy flush_metrics().
