+++ src/vmm/src/devices/legacy/mod.rs
@@ -70,5 +70,6 @@ pub fn flush_metrics<S: Serializer>(serializer: S) -> Result<S::Ok, S::Error> {
     seq.serialize_entry("i8042", &i8042::METRICS)?;
     #[cfg(target_arch = "aarch64")]
     seq.serialize_entry("rtc", &rtc_pl031::METRICS)?;
+    seq.serialize_entry("uart", &serial::METRICS)?;
     seq.end()
 }
